# Sort and unsort the input texts for transformers