Sort the input text by their length so that the padding for each batch can be as small as possible.