@JiushengChen 2020-12-03
Update doc and benchmark.